requirements.txt
requests==2.31.0
flask==3.0.2
gunicorn==21.2.0
python-dotenv==1.0.1
tabulate==0.9.0
flask-cors==4.0.0
flask-limiter==3.5.0
redis==5.0.1
cryptography==41.0.7 